Initial implementation of the API clients
This creates simple clients for the student and studies APIs and
creates types and translations for various things.

This is far from complete, and there is currently no way to get
the data to the main bundle.

This is mainly to have CLI commands for debugging and to have a
proper bundle config, so the bundle can be integrated.