json: expose ISO 3166-1 alpha-3 codes if possible
For countries where we have to code expose it in the json, so we can
for example use it as a shorthand in the UI when there isn't enough space.
Note that not all countries/nationalities in CO can be mapped to a code,
in which case it will be null. Use Kosovo in the test example to test this
edge case.