sync: in case to many students change, fall back to a full sync
Since we now communicate to cabinet via SyncResult() if the sync
was a full or or a partial one, we can now also do a full sync
in case only a partial one was requested.

In case there are 10000 changes because the sync was down for too
long, or CO was restarted or at the start of the semester everything
changed in one go, we don't want to fetch thousands of entries from CO.
It would be slow and we would hammer CO.

In that case stop the partial sync and go back to a full sync. Downside
is that we might go back in time for 15 min since we drop the live sync
data that is stored in the cursor, but hopefully this should be a rare
event.

This also removes the partial sync on the first full sync, since we would
have to prevent endless loops there in case CO is broken (or our logic is)
and it's not worth the complexity imo.